requirements.txt
pandas>=1.5
PyQt6>=6.5.0
PyYAML>=6.0
